Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int r;
-    scanf("%i",r);
-    printf("Area: %i", 3.14*(r*r));
+    float r;
+    scanf("%f",r);
+    printf("Area: %f", 3.14*(r*r));
     return 0;
 }
